Shrunk footer, condensed padding above body, condensed table, began conversion to decimal for dkp

+++ Insomnia.Core/Models/AddDkpToRaiders.cs
@@ -3,6 +3,6 @@
     public class AddDkpToRaiders
     {
         public Raider[] Raiders { get; set; }
-        public int DkpToAdd { get; set; }
+        public decimal DkpToAdd { get; set; }
     }
 }
